Add Viewport component
