+++ src/context/index.ts
@@ -1,2 +1,3 @@
 export * from './HeaderRowContext';
 export * from './ViewportColumnsContext';
+export * from './ViewportContext';
